Misc fixes

+++ Src/Processor/Mania/ManiaScore.cpp
@@ -8,7 +8,7 @@ PP_NAMESPACE_BEGIN
 ManiaScore::ManiaScore(
 	s64 scoreId,
 	EGamemode mode,
-	s32 userId,
+	s64 userId,
 	s32 beatmapId,
 	s32 score,
 	s32 maxCombo,
