+++ Src/Processor/Score.cpp
@@ -12,7 +12,7 @@ PP_NAMESPACE_BEGIN
 Score::Score(
 	s64 scoreId,
 	EGamemode mode,
-	s32 userId,
+	s64 userId,
 	s32 beatmapId,
 	s32 score,
 	s32 maxCombo,
